Add reset feature, adjust reset() function & fix scoreboard
Player had to wait for the game to finish for the score to reset.
Reset button added below play buttons to make it convenient for the player
to reset the game when needed.

reset() function did not reset the scoreboard to 0 unless the player played
at least 1 round. Hence adjusted to set scoreboard to 0 when called.

Scoreboard was empty upon page load. Now preset to 0.